+++ addOns/pscanrules/src/test/java/org/zaproxy/zap/extension/pscanrules/CookieHttpOnlyScanRuleUnitTest.java
@@ -125,7 +125,7 @@ class CookieHttpOnlyScanRuleUnitTest extends PassiveScannerTest<CookieHttpOnlySc
         msg.setResponseHeader(
                 "HTTP/1.1 200 OK\r\n"
                         + "Server: Apache-Coyote/1.1\r\n"
-                        + "Set-Cookie: hasatt=test123; Path=/; HttpOnly\r\n"
+                        + "Set-Cookie: hasattr=test123; Path=/; HttpOnly\r\n"
                         + "Set-Cookie: test=123; Path=/;\r\n"
                         + "Content-Type: text/html;charset=ISO-8859-1\r\n"
                         + "Content-Length: "
@@ -296,7 +296,7 @@ class CookieHttpOnlyScanRuleUnitTest extends PassiveScannerTest<CookieHttpOnlySc
         msg.setResponseHeader(
                 "HTTP/1.1 200 OK\r\n"
                         + "Server: Apache-Coyote/1.1\r\n"
-                        + "Set-Cookie: hasatt=test123; expires=Thu, 01-Jan-1970 00:00:00 GMT; Path=/;\r\n"
+                        + "Set-Cookie: hasattr=test123; expires=Thu, 01-Jan-1970 00:00:00 GMT; Path=/;\r\n"
                         + "Set-Cookie: test=123; Path=/;\r\n"
                         + "Content-Type: text/html;charset=ISO-8859-1\r\n"
                         + "Content-Length: "